requirements.txt
pandas
psycopg2==2.9.3
python-dotenv
SQLAlchemy
pymysql
